improved mqtt callback setting
added check to see whether the lib has been started with begin() method. This will be useful later on as well but for now it allows for a simple check to see whether the mqtt object has been initialized so we dont crash if we try to set the mqtt callback before initialization.