added begin() timeout & other fixes
begin() now has a 2 second timeout period so that if a network isnt connectable it wont block forever. Also fixed some small issues in OTA and added a return statement to loop in the event that the parameter check fails it returns false. previously there was no return in that case