changed connected bool to connectionStatus int
changed the way that we know what is connected. used to be a bool that was true only on both wifi & mqtt connection. Now we have 3 states NO_CONNECTION, WIFI_ONLY, and FULL_CONNECTION. the connected var has been changed to _connectionStatus to reflect the non boolean-ness of the var