Added version in the codebase
The system can now tack on a version identifier to the hostname. This makes it easy to to see which version of the lib any given module is running.  (somewhat of a god object addition... but I find it useful). May rework the functions at some point to make it a bit cleaner.