FS bug fix on new file creation
New files were not being created due to a flipped conditional checking for the file already existing. If the file existed then it would make a new one, if it didnt already exist it would not create a new one.  This has been fixed so that if a file does not exist create it and if it does then leave it alone.